+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased][unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+- UI overlay displays on all desktops
+
+### Security
+
 ## [0.2.0] - 2015-04-25
 ### Added
 - Qt UI overlay